+++ internal/domain.go
@@ -1,0 +1,7 @@
+package internal
+
+type Message {
+
+}
+
+
